quick fix

+++ pointers_arrays_strings/100-atoi.c
@@ -15,7 +15,7 @@
 int _atoi(char *str)
 {
 	int sign = 1;
-	long int num = 0;
+	unsigned int num = 0;
 
 	do
 	{
